Add simple recommender service and associated repositories to aggregate ratings and event logs. Also clean up imports

+++ flix_freaks/src/main/scala/common/domain/Rating.scala
@@ -4,6 +4,6 @@ import java.time.Instant
 
 case class Rating(userId: String,
                   movieId: String,
-                  rating: Float,
+                  rating: BigDecimal,
                   timeStamp: Instant,
                   isExplicit: Boolean = true)
